Created more observables and added a toggle button
I have created several observables for sticky broadcasts. I'm
observing for several system-wide events, such as when the
battery level changes, the device enters airplane mode, every
time a minute passes, and when the screen is on/off. I then
subscribe to all of the observables and log their actions.

In addition, I've implemented a toggle for the subscription.
The first time the button is clicked, we subscribe to the
observables. However, the second time it's clicked, we
unsubscribe.

Overall, nothing overly fancy going on here.