Added a TextView and Clear button
I've added a TextView which displays the broadcasts that are
being logged by the app. There are also now 2 buttons:

Subscribe button - toggles the subscription on/off
Clear button - clears previous broadcasts. This button is disabled
when we are subscribed.

I'm also using a StringBuilder to efficiently concatenate the
Strings that are emitted.